Refactor: Pull out a HasCreator interface
So that `Composite` and the not-actually-a-`type[Node]` `Macro` can avoid duplicating code
